refactor: extract htmlToText to shared utility
