+++ src/androidx/release-notes-parser.ts
@@ -1,34 +1,6 @@
+import { htmlToText } from "../html/to-text.js";
+
 const VERSION_HEADING_RE = /<h[23][^>]*>\s*Version\s+([\d][^\s<]*)\s*<\/h[23]>/gi;
-
-function unescapeEntities(text: string): string {
-  return text
-    .replace(/&lt;/g, "<")
-    .replace(/&gt;/g, ">")
-    .replace(/&quot;/g, '"')
-    .replace(/&#39;/g, "'")
-    .replace(/&amp;/g, "&");
-}
-
-function stripTags(text: string): string {
-  let result = text;
-  let prev: string;
-  do {
-    prev = result;
-    result = result.replace(/<[^>]*>/g, "");
-  } while (result !== prev);
-  return result;
-}
-
-function htmlToText(html: string): string {
-  const formatted = html
-    .replace(/<li[^>]*>/gi, "- ")
-    .replace(/<\/li>/gi, "\n")
-    .replace(/<br\s*\/?>/gi, "\n")
-    .replace(/<\/p>/gi, "\n\n");
-  return unescapeEntities(stripTags(formatted))
-    .replace(/\n{3,}/g, "\n\n")
-    .trim();
-}
 
 export function parseAndroidXReleaseNotes(html: string): Map<string, string> {
   const sections = new Map<string, string>();
